Yeti: LICENSE, README, MODIFICATIONS files
This is the yeti version, added 3rd party dependencies to RenderDoc's
LICENSE.md file. The added deps match what is found in
docs/credits_acknowledgements.rst with the following exceptions because
they are unused in our distribution:

SourceCodePro Fonts - Used by Qt
Qt - Not yet used for the Windows UI
plthook - Used on linux to hook games

Also the licenses for TreeView and IconSet only require that the URI to the
license is included. For the other libraries I appended the full licenses.

Bug: 36366008
Test: None
Change-Id: Ib795da4db059bddc573995f9663e8cc2aa3518ba