Yeti: Work around for one of the RenderDoc checks
RenderDoc has many checks throughout the code and one of them checks
that extension list is sorted in an alphabetical order. It's not clear
why this check is important, so I have changed it to report a warning
and removed the FATAL assert. Also added some TODO(akharlamov) tags to
keep track of unsolved puzzles like that.

Change-Id: Iedc126812efdfa202a314198a32fd5e614b0eb70