Yeti: Workaround for a crash
RenderDoc likes to call std::string str(getenv("HOME")) or something
like that. Unfortunatley on our platform HOME is not defined when
capture_replay is launched and this causes RenderDoc to crash. This CL
adds a workaround to this problem and a small info output to list
paths to renderdoc layer.

Test: launch 3dmark via game_client as described in the bug.
Bug:67605219
Change-Id: I41f82d2a173a607b6460e84c02d917f767ac16f5