Add a script to build code generated project using Ninja on Windows.
The Visual Studio build for code generated project is super slow and doesn't scale well on gWindows with cores. This is an especially sad waste of time if you don't plan to use Visual Studio to debug the project.
This script uses Ninja + VS2015 to build release binaries super fast.

Change-Id: I20d86173b22a6c98757a9444fa8e2ff9a6cd0cbf