Allow to not generate renderpass images in gold_reference shim
gold_reference shim is now compiled twice, once as before to produce all
the intermediate buffers and one to create a simple "screenshot" shim
that outputs only the final swapchain images to be used in automated
tests to validate correctness of rendering.

Added a conditional flag and added another target to produce the
original gold_reference shim and the screenshot shim.

Change-Id: Iaa36f66d01257540944978e937be231eab930eb5