Fix Yeti build of codegen project under Windows Subsystem for Linux
rdoc_auto_capture_shim had '#include <log.h>' which fails to compile for
me. Also there is a race condition when multiple build targets for shims
try to copy libshim_vulkan.so into the working directory. When two or
more do this in parallel, one will create the output file, while the
others will fail to open it and the build fails. Removed the copy build
rule from all but sample_shim.

Change-Id: Ic5c8c454beb11ded285a6ccdee314b4ea8635b34