Remove compiler warnings in CodeGen projects on Windows
There were a number of warnings on windows for data conversions (e.g.
assigning size_t to uint32_t and uint64_t to double) as well as secure
CRT warnings for functions getenv(), fopen(), sprintf(). Now we are
using a secure snprintf() instead and have helpers for getenv() and
fopen() which use _dupenv_s() and fopen_s() on Win32.

Change-Id: I7e91c729fe425ae646447f89ffe59b5d937a29d2