Increase timeout time.
It seems that developers occasionally experience renderdoc remote
server disconnect. I am not able to reproduce this on the traces
attached to the bug. One suspicion is that it may be caused by socket
timeout, which was set to 0.75s. Tentatively increase this to 90s to
see if developers report a better experience.

Bug: 116700822

Change-Id: Icfe00dfe8ddde27d41be81f119781c2b4fa55a12