Fix the code gen output when image memory reset optimization is disabled.
One of the code gen reset optimizations is to avoid reseting the memory backing
images, because the images need to be reset as images, not memory. This can be
disabled for debugging by setting `RDOC_CODE_GEN_OPT_IMAGE_MEMORY=false` or
`RDOC_CODE_GEN_ALL_OPTS=false`.

This patch changes the code gen output when this optimization is disabled, so
that the backing memory for each image will be reset before each frame;
prior to this patch, disabling the optimization would only initialize the
backing memory once.

These extra resets of the image backing memory are safe, assuming that device
memory initial content chunks always precede image initial content chunks. This
seems to be the case currently, and if it isn't the case, it could cause bugs
even when the optimization is on (default), because the memory may be forced to
reset due to aliasing. I will add a separate CL to assert that this is the case.

Change-Id: Ie35afb83df8ae9518b6f2baccbd9f3186f4239e2