Add a batch file to compile the test package from an .rdc file
This batch file takes an .rdc file and produces all the necessary data
files for automated tests. This includes a compressed archive containing
the compiled codegen binary and associated data files for all three
platforms (Yeti, Windows, Linux), reference images for screenshot and
rdoc_auto_capture tests, reference thumbnail and original .rdc files in
respective subdirectories of the output folder.

To run:

  generate_test_package.cmd <rdc_file> <out_dir> [<package>]

The optional "package" parameter specifies the output name of the
package (used for archive, screenshot, etc) names. If not specified, the
name of .rdc file is used as package name.

Change-Id: I3b72b6ae4b08e2e52bd85e52bbe0e4e7971143d2