Change CodeGen to pass 'aux' by pointer/reference instead of value
When calling helper functions like ImageLayoutTransition(), the 'aux'
parameter was passed by value and a copy constructor was called on a
stack-allocated copy of the structure. Since our clang doesn't support
stack 'coloring' technique to reuse stack allocations of variables
inside different scopes, this all compounded to using new copy of the
stack variable for each invocation which, for some traces, can easily
blow up the stack.

Changed this to use const reference and pointers to not do the
unnecessary copies.

Bug: 117420665
Test: 1.19_inf_1153

Change-Id: Ie458d87cadc004b4ae88d8f6bb3fdee188d7e60f