Make capture possible with a NULL window.
When an application renders into a virtual swapchain,
it can specify a NULL window handle. In that case RenderDoc
will fail to capture such an app, because certain areas of
the code assume window handle to be != NULL. This CL fixes
it to allow RenderDoc to capture over NULL windows.

Change-Id: If0cf0c67da81811340a89edad90de4cfcaf313dd