Rebranding yeti->ggp
Support yeti/ggp capture, and ggp only replay/codegen.
yeti.cmake was removed from windows SDK, so I removed the build_yeti.sh
for codegen, however I still left the support for __yeti__ and yeti
surface in codegen, so it is technically possible to generate yeti
app if you have yeti.cmake and adjust the build script.

ggp.cmake is not in the current base SDK, but that will be added,
and I will be able to switch to the official one, for now using a
custom one created by sfunkenhauser@, but will change before submission.

BUG: b/123072536

Change-Id: I3fd98b99deb16cf393f4130e52c5db2530e6f22b