Vulkan: Optimize image resets for Vulkan replay
This uses the new image tracking to avoid unnecessary copying of image
data when applying the initial state.

This does not attempt to optimize the layout transitions needed for
these image copies, and also does not attempt to optimize images whose
initial contents are `ClearColorImage` or `CleardepthStencilImage`, nor
does this optimize the reset of multisample images.

Change-Id: I449eabfe969229fa64e233760aaadb489776ee23